+++ packages/eyes-sdk-core/lib/utils/getBrowserInfo.js
@@ -3,7 +3,9 @@
 const TypeUtils = require('./TypeUtils')
 
 async function getBrowserInfo({browser, getEmulatedDevicesSizes, getIosDevicesSizes}) {
-  if (TypeUtils.has(browser, 'name')) {
+  const isMobile =
+    browser.deviceName || browser.mobile || browser.iosDeviceInfo || browser.chromeEmulationInfo
+  if (!isMobile) {
     const {name, width, height} = browser
     return {name, width, height}
   } else {